Move all framework components into fw dir; rewrite imports of consumers
